// management-system-frontend/src/utils/Enums.tsx
export enum RequestMethod {
  GET = "GET",
  POST = "POST",
  DELETE = "DELETE",
  PUT = "PUT",
  PATCH = "PATCH",
}

export enum ToastType {
  SUCCESS = "success",
  ERROR = "error",
  WARNING = "warning",
  INFO = "info",
  DEFAULT = "default",
}

export enum AUTH_METHOD {
  LOGIN = 0,
  SIGNUP = 1,
}

export enum PriorityType {
  LOW = 0,
  MID = 1,
  HIGH = 2,
}
export enum TaskStatus {
  TODO = 0,
  IN_PROGRESS = 1,
  DONE = 2,
}
export enum CrudOperation {
  CREATE = "CREATE",
  UPDATE = "UPDATE",
  DELETE = "DELETE",
}